Factoring Guardian Weekly functionality specific code out of Processor
